closing context menu on click outside, code cleanup, removed redundend code.
