+++ src/app/calender-body/month-body/month-body.component.html
@@ -10,7 +10,7 @@
     </div>
 </div>
 
-<div id="contextMenu" [ngStyle]="rightPanelStyle" class="context-menu" (clickOutSide)="closeContextMenu()">
+<div id="contextMenu" [ngStyle]="rightPanelStyle" class="context-menu">
     <ul class="menu">
         <li><a (click)="editEvent()"><i class="bi bi-pencil-square" area-hidden="true"></i>Edit</a></li>
         <hr>
