+++ src/app/calender-body/month-body/month-body.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, OnDestroy, OnInit, ViewChild } from '@angular/core';
+import { Component, Input, OnDestroy, OnInit, ViewChild, Renderer2 } from '@angular/core';
 import { MonthBodyService } from './month-body.service';
 import { CalenderBodyService } from '../calender-body.service';
 import { ChangeMonthService } from '../change-month/change-month.service';
@@ -38,7 +38,8 @@ export class MonthBodyComponent implements OnInit,OnDestroy {
     private readonly monthBodyService: MonthBodyService,
     private readonly changeMonthService: ChangeMonthService,
     private readonly calenderBodyService: CalenderBodyService,
-    private readonly headerService: HeaderService
+    private readonly headerService: HeaderService,
+    private renderer: Renderer2
   ) {}
 
   ngOnInit(): void {
@@ -112,9 +113,13 @@ export class MonthBodyComponent implements OnInit,OnDestroy {
         }
       })
     );
+      
+    this.renderer.listen('window', 'click',(e:Event)=>{ 
+      this.closeContextMenu();
+    });
 
     this.closeContextMenu();
-  }
+  };
 
   detectRightMouseClick(
     $event: MouseEvent,
@@ -132,24 +137,24 @@ export class MonthBodyComponent implements OnInit,OnDestroy {
         'top.px': $event.clientY - 90,
       };
     }
-  }
+  };
 
   closeContextMenu() {
     this.rightPanelStyle = {
       display: 'none',
     };
-  }
+  };
 
   openEventDialog(): void {
     const evenDialog = document.getElementById('editEventModal');
     if (evenDialog) {
       evenDialog.style.display = 'block';
     }
-  }
+  };
 
   onSubmit(form: NgForm): void {
-    this.isDateValid = this.validateDate(form.value.date);
-    this.isTitleValid = this.validateTitle(form.value.title);
+    this.isDateValid = this.headerService.validateDate(form.value.date);
+    this.isTitleValid = this.headerService.validateTitle(form.value.title);
     if (this.isDateValid && this.isTitleValid) {
       for (const [i, day] of this.daysObj.entries()) {
         let dateArr = this.event.date.split('/');
@@ -169,64 +174,18 @@ export class MonthBodyComponent implements OnInit,OnDestroy {
     } else if (!this.isDateValid) {
       alert('Please enter date in dd/mm/yyyy format!');
     }
-  }
-
-  validateTitle(title: string): boolean {
-    if (!title) {
-      return false;
-    }
-    return true;
-  }
-
-  validateDate(date: string): boolean {
-    let dateformat = /^(0?[1-9]|[1-2][0-9]|3[01])[\/](0?[1-9]|1[0-2])/;
-
-    if (date.match(dateformat)) {
-      let operator = date.split('/');
-      let datepart: any[] = [];
-
-      if (operator.length > 1) {
-        datepart = date.split('/');
-      }
-
-      let day = parseInt(datepart[0]);
-      let month = parseInt(datepart[1]);
-      let year = parseInt(datepart[2]);
-      let ListofDays = [31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31];
-
-      if (month > 12) {
-        return false;
-      }
-
-      if (month == 1 || month > 2) {
-        if (day > ListofDays[month - 1]) {
-          return false;
-        }
-      } else if (month == 2) {
-        let leapYear = false;
-        if ((!(year % 4) && year % 100) || !(year % 400)) leapYear = true;
-        if (leapYear == false && day >= 29) {
-          return false;
-        } else if (leapYear == true && day > 29) {
-          return false;
-        }
-      }
-    } else {
-      return false;
-    }
-    return true;
-  }
+  };
 
   closeEventDialog(): void {
     const evenDialog = document.getElementById('editEventModal');
     if (evenDialog) {
       evenDialog.style.display = 'none';
     }
-  }
+  };
 
   resetForm(): void {
     this.editEventForm.reset();
-  }
+  };
 
   editEvent() {
     this.editEventForm.setValue({
@@ -234,7 +193,8 @@ export class MonthBodyComponent implements OnInit,OnDestroy {
       date: this.event.date,
     });
     this.openEventDialog();
-  }
+    this.closeContextMenu();
+  };
 
   deleteEvent() {
     for (const [i, day] of this.daysObj.entries()) {
@@ -246,7 +206,7 @@ export class MonthBodyComponent implements OnInit,OnDestroy {
     }
     this.calenderBodyService.deleteEvent(this.eventId);
     this.closeContextMenu();
-  }
+  };
 
   drop(event: CdkDragDrop<{ id: string; title: string; date: string }[]>) {
     let draggedEvent = event.previousContainer.data[event.currentIndex];
@@ -261,11 +221,11 @@ export class MonthBodyComponent implements OnInit,OnDestroy {
       }
     }
     this.calenderBodyService.changeEventDate(draggedEvent, dropDate);
-  }
+  };
 
   ngOnDestroy(): void {
     this.subscriptions.unsubscribe();
-  }
+  };
   // let dateArr = [
   //   {dateId: '', date: new Date(), events: [{id: '', title: '', date: ''}]}
   // ]
